リクエストのURLの部分を見るようにしました 多分できていると思います close #31, close #23, close #34

+++ src/file_opener.cpp
@@ -13,7 +13,7 @@
 
 std::vector<std::string> lineSpliter(std::string origin, std::string delim);
 
-static size_t setFileSize(std::string Filename, Response &res) {	
+static unsigned long long setFileSize(std::string Filename, Response &res) {	
 	unsigned long long		bytes = 0;
 	int						status = 0;
 	int						fd;
